Refactor CodeLine State Management
