+++ src/dotnet/APIView/APIViewWeb/Client/src/pages/review.ts
@@ -1,293 +1,24 @@
-import Split from "split.js";
-import { updatePageSettings, toggleCommentIcon } from "../shared/helpers";
+
 import { rightOffCanvasNavToggle } from "../shared/off-canvas";
 
+import * as rvM from "./review.module"
+import * as hp from "../shared/helpers";
+
 $(() => {  
-  const SEL_DOC_CLASS = ".documentation";
-  const SHOW_DOC_CHECK_COMPONENT = "#show-documentation-component";
   const SHOW_DOC_CHECKBOX = ".show-documentation-checkbox";
   const SHOW_DOC_HREF = ".show-documentation-switch";
   const SHOW_DIFFONLY_CHECKBOX = ".show-diffonly-checkbox";
   const SHOW_DIFFONLY_HREF = ".show-diffonly-switch";
   const TOGGLE_DOCUMENTATION = ".line-toggle-documentation-button";
   const SEL_HIDDEN_CLASS = ".hidden-api-toggleable";
-  const SHOW_HIDDEN_CHECK_COMPONENT = "#show-hidden-api-component";
   const SHOW_HIDDEN_CHECKBOX = "#show-hidden-api-checkbox";
   const SHOW_HIDDEN_HREF = ".show-hidden-api";
 
-  hideCheckboxesIfNotApplicable();
+  rvM.hideCheckboxesIfNotApplicable();
 
   /* FUNCTIONS
   --------------------------------------------------------------------------------------------------------------------------------------------------------*/
-  function hideCheckboxesIfNotApplicable() {
-    if ($(SEL_DOC_CLASS).length == 0) {
-      $(SHOW_DOC_CHECK_COMPONENT).hide();
-    }
-    if ($(SEL_HIDDEN_CLASS).length == 0) {
-      $(SHOW_HIDDEN_CHECK_COMPONENT).hide();
-    }
-  }
 
-  /* Split left and right review panes using split.js */
-  function splitReviewPageContent() {
-    const rl = $('#review-left');
-    const rr = $('#review-right');
-
-    if (rl.length && rr.length) {
-      Split(['#review-left', '#review-right'], {
-        direction: 'horizontal',
-        sizes: [17, 83],
-        elementStyle: (dimension, size, gutterSize) => {
-          return {
-            'flex-basis': `calc(${size}% - ${gutterSize}px`
-          }
-        },
-        gutterStyle: (dimension, gutterSize) => {
-          return {
-            'flex-basis': `${gutterSize}px`
-          }
-        }
-      });
-    }
-  }
-
-  /* Update Icons that indicate if Section is Expanded or Collapsed */
-  function updateSectionHeadingIcons(setTo: string, caretIcon, headingRow) {
-    if (setTo == "OPEN") {
-      caretIcon.removeClass("fa-angle-right");
-      caretIcon.addClass("fa-angle-down");
-      headingRow.find(".row-fold-elipsis").addClass("d-none");
-    }
-
-    if (setTo == "CLOSE") {
-      caretIcon.removeClass("fa-angle-down");
-      caretIcon.addClass("fa-angle-right");
-      headingRow.find(".row-fold-elipsis").removeClass("d-none");
-    }
-  }
-
-  /* Expand or Collapse CodeLine Top Level Sections */
-  function toggleSectionContent(headingRow, sectionContent, caretDirection, caretIcon) {
-    if (caretDirection.endsWith("right")) {
-      // In case the section passed has already been replaced with more rows
-      if (sectionContent.length == 1) {
-        const sectionContentClass = sectionContent[0].className.replace(/\s/g, '.');
-        const sectionCommentClass = sectionContentClass.replace("code-line.", "comment-row.");
-        sectionContent = $(`.${sectionContentClass}`);
-        sectionContent.push(...$(`.${sectionCommentClass}`));
-      }
-
-      $.each(sectionContent, function (index, value) {
-        let rowClasses = $(value).attr("class");
-        if (rowClasses) {
-          if (rowClasses.match(/lvl_1_/)) {
-            if (rowClasses.match(/comment-row/) && !$("#show-comments-checkbox").prop("checked")) {
-              toggleCommentIcon($(value).attr("data-line-id"), true);
-              return; // Dont show comment row if show comments setting is unchecked
-            }
-            disableCommentsOnInRowTables($(value));
-            $(value).removeClass("d-none");
-            $(value).find("svg").attr("height", `${$(value).height()}`);
-          }
-        }
-      });
-
-      // Update section heading icons to open state
-      updateSectionHeadingIcons("OPEN", caretIcon, headingRow);
-    }
-    else {
-      $.each(sectionContent, function (index, value) {
-        let rowClasses = $(value).attr("class");
-        if (rowClasses) {
-          if (rowClasses.match(/lvl_[0-9]+_parent_/)) {
-            // Update all heading/parent rows to closed state before hiding it
-            let caretIcon = $(value).find(".row-fold-caret").children("i");
-            updateSectionHeadingIcons("CLOSE", caretIcon, $(value));
-          }
-        }
-        $(value).addClass("d-none");
-      });
-
-      // Update section heading icons to closed state
-      updateSectionHeadingIcons("CLOSE", caretIcon, headingRow);
-    }
-  }
-
-  /* Expand or Collapse CodeLine SubSections */
-  function toggleSubSectionContent(headingRow, subSectionLevel, subSectionHeadingPosition, subSectionContentClass, caretDirection, caretIcon, lineNumber) {
-    var subSectionDescendants = $(`.${subSectionContentClass}`);
-
-    if (caretDirection.endsWith("right")) {
-      var startShowing = false;
-
-      $.each(subSectionDescendants, function (index, value) {
-        var rowClasses = $(value).attr("class");
-        var rowLineNumber = $(value).find(".line-number>span").text();
-        if (rowClasses) {
-          if (rowClasses.match(new RegExp(`lvl_${subSectionLevel}_parent_${subSectionHeadingPosition}`)) && rowLineNumber == lineNumber)
-            startShowing = true;
-
-          if (startShowing && (rowClasses.match(new RegExp(`lvl_${subSectionLevel}_parent_${Number(subSectionHeadingPosition) + 1}`))
-            || rowClasses.match(new RegExp(`lvl_${subSectionLevel}_child_${Number(subSectionHeadingPosition) + 1}`))
-            || rowClasses.match(new RegExp(`lvl_${Number(subSectionLevel) - 1}_`))))
-            return false;
-
-          // Show only immediate descendants
-          if (startShowing) {
-            if (rowClasses.match(new RegExp(`lvl_${Number(subSectionLevel) + 1}_`))) {
-              if (rowClasses.match(/comment-row/) && !$("#show-comments-checkbox").prop("checked")) {
-                toggleCommentIcon($(value).attr("data-line-id"), true);
-                return; // Dont show comment row if show comments setting is unchecked
-              }
-
-              disableCommentsOnInRowTables($(value));
-              $(value).removeClass("d-none");
-              let rowHeight = $(value).height() ?? 0;
-              $(value).find("svg").attr("height", `${rowHeight}`);
-            }
-          }
-        }
-      });
-
-      // Update section heading icons to open state
-      updateSectionHeadingIcons("OPEN", caretIcon, headingRow);
-    }
-    else {
-      var startHiding = false;
-
-      $.each(subSectionDescendants, function (index, value) {
-        var rowClasses = $(value).attr("class");
-        var rowLineNumber = $(value).find(".line-number>span").text();
-        if (rowClasses) {
-          if (rowClasses.match(new RegExp(`lvl_${subSectionLevel}_parent_${subSectionHeadingPosition}`)) && rowLineNumber == lineNumber)
-            startHiding = true;
-
-          if (startHiding && (rowClasses.match(new RegExp(`lvl_${subSectionLevel}_parent_${Number(subSectionHeadingPosition) + 1}`))
-            || rowClasses.match(new RegExp(`lvl_${subSectionLevel}_child_${Number(subSectionHeadingPosition) + 1}`))
-            || rowClasses.match(new RegExp(`lvl_${Number(subSectionLevel) - 1}_`))))
-            return false;
-
-          if (startHiding) {
-            let descendantClasses = rowClasses.split(' ').filter(c => c.match(/lvl_[0-9]+_child_.*/))[0];
-            if (descendantClasses) {
-              let descendantLevel = descendantClasses.split('_')[1];
-              if (/^\d+$/.test(descendantLevel)) {
-                if (Number(descendantLevel) > Number(subSectionLevel)) {
-                  $(value).addClass("d-none");
-                  if (rowClasses.match(/lvl_[0-9]+_parent_.*/)) {
-                    // Update all heading/parent rows to closed state before hiding it
-                    let caretIcon = $(value).find(".row-fold-caret").children("i");
-                    updateSectionHeadingIcons("CLOSE", caretIcon, $(value));
-
-                  }
-                }
-              }
-            }
-          }
-        }
-      });
-
-      // Update section heading icons to closed state
-      updateSectionHeadingIcons("CLOSE", caretIcon, headingRow);
-    }
-  }
-
-  /* On Click Handler for Expand/Collapse of CodeLine Sections and SubSections  */
-  function toggleCodeLines(headingRow) {
-    if (headingRow.attr('class')) {
-      const headingRowClasses = headingRow.attr('class').split(/\s+/);
-      const caretIcon = headingRow.find(".row-fold-caret").children("i");
-      const caretDirection = caretIcon.attr("class").split(/\s+/).filter(c => c.startsWith('fa-angle-'))[0];
-      const subSectionHeadingClass = headingRowClasses.filter(c => c.startsWith('code-line-section-heading-'))[0];
-      const subSectionContentClass = headingRowClasses.filter(c => c.startsWith('code-line-section-content-'))[0];
-
-      if (subSectionHeadingClass) {
-        const sectionKey = subSectionHeadingClass.replace("code-line-section-heading-", "")
-        const sectionKeyA = headingRowClasses.filter(c => c.startsWith('rev-a-heading-'))[0]?.replace('rev-a-heading-', '');
-        const sectionKeyB = headingRowClasses.filter(c => c.startsWith('rev-b-heading-'))[0]?.replace('rev-b-heading-', '');
-
-        if (/^\d+$/.test(sectionKey)) {
-          var sectionContent = $(`.code-line-section-content-${sectionKey}`);
-          if (sectionContent.hasClass("section-loaded")) {
-            toggleSectionContent(headingRow, sectionContent, caretDirection, caretIcon);
-          }
-          else {
-            let uri = '?handler=codelinesection';
-            const uriPath = location.pathname.split('/');
-            const reviewId = uriPath[uriPath.length - 1];
-            const revisionId = new URLSearchParams(location.search).get("revisionId");
-            const diffRevisionId = new URLSearchParams(location.search).get("diffRevisionId");
-            const diffOnly = new URLSearchParams(location.search).get("diffOnly");
-            uri = uri + '&id=' + reviewId + '&sectionKey=' + sectionKey;
-            if (revisionId)
-              uri = uri + '&revisionId=' + revisionId;
-            if (diffRevisionId)
-              uri = uri + '&diffRevisionId=' + diffRevisionId;
-            if (diffOnly)
-              uri = uri + '&diffOnly=' + diffOnly;
-            if (sectionKeyA)
-              uri = uri + '&sectionKeyA=' + sectionKeyA;
-            if (sectionKeyB)
-              uri = uri + '&sectionKeyB=' + sectionKeyB;
-
-            const loadingMarkUp = "<td class='spinner-border spinner-border-sm ms-4' role='status'><span class='sr-only'>Loading...</span></td>";
-            const failedToLoadMarkUp = "<div class='alert alert-warning alert-dismissible fade show' role='alert'>Failed to load section. Refresh page and try again.</div>";
-            if (sectionContent.children(".spinner-border").length == 0) {
-              sectionContent.children("td").after(loadingMarkUp);
-            }
-            sectionContent.removeClass("d-none");
-
-            const request = $.ajax({ url: uri });
-            request.done(function (partialViewResult) {
-              sectionContent.replaceWith(partialViewResult);
-              toggleSectionContent(headingRow, sectionContent, caretDirection, caretIcon);
-              addToggleEventHandlers();
-            });
-            request.fail(function () {
-              if (sectionContent.children(".alert").length == 0) {
-                sectionContent.children(".spinner-border").replaceWith(failedToLoadMarkUp);
-              }
-            });
-            return request;
-          }
-        }
-      }
-
-      if (subSectionContentClass) {
-        const subSectionClass = headingRowClasses.filter(c => c.match(/.*lvl_[0-9]+_parent.*/))[0];
-        const lineNumber = headingRow.find(".line-number>span").text();
-        if (subSectionClass) {
-          const subSectionLevel = subSectionClass.split('_')[1];
-          const subSectionHeadingPosition = subSectionClass.split('_')[3];
-          if (/^\d+$/.test(subSectionLevel) && /^\d+$/.test(subSectionHeadingPosition)) {
-            toggleSubSectionContent(headingRow, subSectionLevel, subSectionHeadingPosition, subSectionContentClass, caretDirection, caretIcon, lineNumber);
-          }
-        }
-      }
-    }
-  }
-
-  /* Add event handler for Expand/Collapse of CodeLine Sections and SubSections */
-  function addToggleEventHandlers() {
-    $('.row-fold-elipsis, .row-fold-caret').on('click', function (event) {
-      event.preventDefault();
-      event.stopImmediatePropagation();
-      var headingRow = $(event.currentTarget).parents('.code-line').first();
-      toggleCodeLines(headingRow);
-
-    });
-  }
-
-  /* Disables Comments for tables within codeline rows. Used for code-removed lines in diff */
-  function disableCommentsOnInRowTables(row: JQuery<HTMLElement>) {
-    if (row.hasClass("code-removed")) {
-      const innerTable = row.find(".code-inner>table");
-      if (innerTable.length > 0) {
-        innerTable.find("tr").removeAttr("data-inline-id");
-        innerTable.find(".line-comment-button").remove();
-      }
-    }
-  }
 
   // Enable SumoSelect
   $(document).ready(function () {
@@ -314,7 +45,7 @@ $(() => {
             var caretClasses = caretIcon.attr("class");
             var caretDirection = caretClasses ? caretClasses.split(' ').filter(c => c.startsWith('fa-angle-'))[0] : "";
             if (caretDirection.endsWith("right")) {
-              $.when(toggleCodeLines(targetAnchorRow)).then(function () {
+              $.when(rvM.toggleCodeLines(targetAnchorRow)).then(function () {
                 navItemRow.removeClass("nav-list-collapsed");
               });
             }
@@ -335,7 +66,7 @@ $(() => {
   if (!$("#review-left").hasClass("d-none"))
   {
     // Only Add Split gutter if left navigation is not hidden
-    splitReviewPageContent();
+    rvM.splitReviewPageContent();
   }
 
   /* TOGGLE PAGE OPTIONS
@@ -362,7 +93,7 @@ $(() => {
   });
 
   $(SHOW_HIDDEN_CHECKBOX).on("click", e => {
-    updatePageSettings(function() {
+    hp.updatePageSettings(function() {
       $(SEL_HIDDEN_CLASS).toggleClass("d-none");
     });
   });
@@ -376,13 +107,13 @@ $(() => {
   });
 
   $("#hide-line-numbers").on("click", e => {
-    updatePageSettings(function(){
+    hp.updatePageSettings(function(){
       $(".line-number").toggleClass("d-none");
     });
   });
 
   $("#hide-left-navigation").on("click", e => {
-    updatePageSettings(function(){
+    hp.updatePageSettings(function(){
       var leftContainer = $("#review-left");
       var rightContainer = $("#review-right");
       var gutter = $(".gutter-horizontal");
@@ -391,7 +122,7 @@ $(() => {
         leftContainer.removeClass("d-none");
         rightContainer.removeClass("col-12");
         rightContainer.addClass("col-10");
-        splitReviewPageContent();
+        rvM.splitReviewPageContent();
       }
       else {
         leftContainer.addClass("d-none");
@@ -474,13 +205,13 @@ $(() => {
 
   /* COLLAPSIBLE CODE LINES (EXPAND AND COLLAPSE FEATURE)
   --------------------------------------------------------------------------------------------------------------------------------------------------------*/
-  addToggleEventHandlers();
+  rvM.addToggleEventHandlers();
 
   /* RIGHT OFFCANVAS OPERATIONS
   --------------------------------------------------------------------------------------------------------------------------------------------------------*/
    // Open / Close right Offcanvas Menu
   $("#review-right-offcanvas-toggle").on('click', function () {
-    updatePageSettings(function () {
+    hp.updatePageSettings(function () {
       rightOffCanvasNavToggle("review-main-container");
     });
   });
