+++ src/dotnet/APIView/APIViewWeb/Client/tests/unit/shared/helpers.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import * as hp from "../../src/shared/helpers";
+import * as hp from "../../../src/shared/helpers";
 
 test('getCookies return valid single cookie', async ({ page }) => {
   const testCookieString = 'preferred_color_mode=light; tz=America%2FLos_Angeles; _octo=GH1.1.1383888135.1683830525'
